pyproject.toml
[project]
name = "streamlitcrewai"
version = "0.1.0"
description = "Creador de Historias Visuales con CrewAI para Redes Sociales"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "streamlit>=1.28.0",
    "crewai>=0.28.0",
    "python-dotenv>=1.0.0",
    "supabase>=2.0.0",
    "transformers>=4.35.0",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "Pillow>=10.0.0",
    "reportlab>=4.0.0",
    "markdown>=3.5.0",
    "agentops>=0.2.0"
]
